+++ src/charts/ColumnChart/ColumnChart.js
@@ -112,7 +112,7 @@ class ColumnChart extends React.PureComponent {
   };
 
   render() {
-    const { data, title, subtitle, allowUpdate, theme = defaultTheme } = this.props;
+    const { data, title, subtitle, allowUpdate, theme = defaultTheme, height = null } = this.props;
     if (!data) {
       return null;
     }
@@ -133,6 +133,7 @@ class ColumnChart extends React.PureComponent {
         backgroundColor: theme.white,
         styledMode: true,
         groupPadding: 0,
+        height,
       },
       plotOptions: {
         series: {
@@ -207,8 +208,6 @@ class ColumnChart extends React.PureComponent {
 }
 
 ColumnChart.defaultProps = {
-  title: false,
-  subtitle: false,
   allowUpdate: false,
   caption: ({ y }) => y,
 };
@@ -223,6 +222,7 @@ ColumnChart.propTypes = {
     })
   ).isRequired,
   caption: PropTypes.func,
+  height: PropTypes.number,
   /**
    * Allows HighCharts chart update. Use it if you plan to provide new data.
    */